refactor(context): move tests out of module
